added changes

+++ setup.ps1
@@ -9,8 +9,7 @@ function Test-InternetConnection {
     try {
         $testConnection = Test-Connection -ComputerName www.google.com -Count 1 -ErrorAction Stop
         return $true
-    }
-    catch {
+    } catch {
         Write-Warning "Internet connection is required but not available. Please check your connection."
         return $false
     }
@@ -25,35 +24,42 @@ function Install-NerdFonts {
     )
 
     try {
-        [void] [System.Reflection.Assembly]::LoadWithPartialName("System.Drawing")
-        $fontFamilies = (New-Object System.Drawing.Text.InstalledFontCollection).Families.Name
-        if ($fontFamilies -notcontains "${FontDisplayName}") {
-            $fontZipUrl = "https://github.com/ryanoasis/nerd-fonts/releases/download/v${Version}/${FontName}.zip"
-            $zipFilePath = "$env:TEMP\${FontName}.zip"
-            $extractPath = "$env:TEMP\${FontName}"
+        if ($IsWindows) {
+            [void] [System.Reflection.Assembly]::LoadWithPartialName("System.Drawing")
+            $fontFamilies = (New-Object System.Drawing.Text.InstalledFontCollection).Families.Name
+            if ($fontFamilies -notcontains "${FontDisplayName}") {
+                $fontZipUrl = "https://github.com/ryanoasis/nerd-fonts/releases/download/v${Version}/${FontName}.zip"
+                $zipFilePath = "$env:TEMP\${FontName}.zip"
+                $extractPath = "$env:TEMP\${FontName}"
 
-            $webClient = New-Object System.Net.WebClient
-            $webClient.DownloadFileAsync((New-Object System.Uri($fontZipUrl)), $zipFilePath)
+                $webClient = New-Object System.Net.WebClient
+                $webClient.DownloadFileAsync((New-Object System.Uri($fontZipUrl)), $zipFilePath)
 
-            while ($webClient.IsBusy) {
-                Start-Sleep -Seconds 2
-            }
-
-            Expand-Archive -Path $zipFilePath -DestinationPath $extractPath -Force
-            $destination = (New-Object -ComObject Shell.Application).Namespace(0x14)
-            Get-ChildItem -Path $extractPath -Recurse -Filter "*.ttf" | ForEach-Object {
-                If (-not(Test-Path "C:\Windows\Fonts\$($_.Name)")) {
-                    $destination.CopyHere($_.FullName, 0x10)
+                while ($webClient.IsBusy) {
+                    Start-Sleep -Seconds 2
                 }
-            }
 
-            Remove-Item -Path $extractPath -Recurse -Force
-            Remove-Item -Path $zipFilePath -Force
-        } else {
-            Write-Host "Font ${FontDisplayName} already installed"
+                Expand-Archive -Path $zipFilePath -DestinationPath $extractPath -Force
+                $destination = (New-Object -ComObject Shell.Application).Namespace(0x14)
+                Get-ChildItem -Path $extractPath -Recurse -Filter "*.ttf" | ForEach-Object {
+                    If (-not(Test-Path "C:\Windows\Fonts\$($_.Name)")) {
+                        $destination.CopyHere($_.FullName, 0x10)
+                    }
+                }
+
+                Remove-Item -Path $extractPath -Recurse -Force
+                Remove-Item -Path $zipFilePath -Force
+            } else {
+                Write-Host "Font ${FontDisplayName} already installed"
+            }
+        } elseif ($IsMacOS) {
+          
+            brew tap homebrew/cask-fonts
+            brew install --cask font-cascadia-code-nerd-font
+            Write-Host "✅ font-cascadia-code-nerd-font installed." -ForegroundColor Green
+    
         }
-    }
-    catch {
+    } catch {
         Write-Error "Failed to download or install ${FontDisplayName} font. Error: $_"
     }
 }
@@ -70,8 +76,7 @@ if (!(Test-Path -Path $PROFILE -PathType Leaf)) {
         $profilePath = ""
         if ($PSVersionTable.PSEdition -eq "Core") {
             $profilePath = "$env:userprofile\Documents\Powershell"
-        }
-        elseif ($PSVersionTable.PSEdition -eq "Desktop") {
+        } elseif ($PSVersionTable.PSEdition -eq "Desktop") {
             $profilePath = "$env:userprofile\Documents\WindowsPowerShell"
         }
 
@@ -82,19 +87,16 @@ if (!(Test-Path -Path $PROFILE -PathType Leaf)) {
         Invoke-RestMethod https://github.com/ChrisTitusTech/powershell-profile/raw/main/Microsoft.PowerShell_profile.ps1 -OutFile $PROFILE
         Write-Host "The profile @ [$PROFILE] has been created."
         Write-Host "If you want to make any personal changes or customizations, please do so at [$profilePath\Profile.ps1] as there is an updater in the installed profile which uses the hash to update the profile and will lead to loss of changes"
-    }
-    catch {
+    } catch {
         Write-Error "Failed to create or update the profile. Error: $_"
     }
-}
-else {
+} else {
     try {
         Get-Item -Path $PROFILE | Move-Item -Destination "oldprofile.ps1" -Force
         Invoke-RestMethod https://github.com/ChrisTitusTech/powershell-profile/raw/main/Microsoft.PowerShell_profile.ps1 -OutFile $PROFILE
         Write-Host "The profile @ [$PROFILE] has been created and old profile removed."
         Write-Host "Please back up any persistent components of your old profile to [$HOME\Documents\PowerShell\Profile.ps1] as there is an updater in the installed profile which uses the hash to update the profile and will lead to loss of changes"
-    }
-    catch {
+    } catch {
         Write-Error "Failed to backup and update the profile. Error: $_"
     }
 }
@@ -102,8 +104,7 @@ else {
 # OMP Install
 try {
     winget install -e --accept-source-agreements --accept-package-agreements JanDeDobbeleer.OhMyPosh
-}
-catch {
+} catch {
     Write-Error "Failed to install Oh My Posh. Error: $_"
 }
 
@@ -119,24 +120,21 @@ if ((Test-Path -Path $PROFILE) -and (winget list --name "OhMyPosh" -e) -and ($fo
 
 # Choco install
 try {
-    Set-ExecutionPolicy Bypass -Scope Process -Force; [System.Net.ServicePointManager]::SecurityProtocol = [System.Net.ServicePointManager]::SecurityProtocol -bor 3072; iex ((New-Object System.Net.WebClient).DownloadString('https://community.chocolatey.org/install.ps1'))
-}
-catch {
+    Set-ExecutionPolicy Bypass -Scope Process -Force; [System.Net.ServicePointManager]::SecurityProtocol = [System.Net.ServicePointManager]::SecurityProtocol -bor 3072; Invoke-Expression ((New-Object System.Net.WebClient).DownloadString('https://community.chocolatey.org/install.ps1'))
+} catch {
     Write-Error "Failed to install Chocolatey. Error: $_"
 }
 
 # Terminal Icons Install
 try {
     Install-Module -Name Terminal-Icons -Repository PSGallery -Force
-}
-catch {
+} catch {
     Write-Error "Failed to install Terminal Icons module. Error: $_"
 }
 # zoxide Install
 try {
     winget install -e --id ajeetdsouza.zoxide
     Write-Host "zoxide installed successfully."
-}
-catch {
+} catch {
     Write-Error "Failed to install zoxide. Error: $_"
 }
